+++ client/src/components/Team/team.css
@@ -1,0 +1,3 @@
+.team-title{
+    font-size: 30px;
+}
